Initial development versions.
This is the inital files used to devlop the doce before it was imported into Github. Version 7a_incBG is the version that was current when the code was imported to GitHub. My idea is that the initial devlopement branch is the one used until the file and function are restructered into a more 'final' arrangement.