Separate Fit Parameter and previous fit files.
I have moved all the settings into a separate file that is called with the main script. There main script also now saves a temporary file with all the fit coefficients in. This is propagated into the next image's fit if present.

There is also a bug fix for the multifit write.