Add variable peak profile and symmetry
The ability to fit for peak profile has been added to the code.

The cahnges included making functions in PeakFourierFunctions for gaussian lorentz and pseudovoigt peaks. All other functions now call these.

There is also the option to fix the peak profile by adding an extra dictionary/library/whatever it is called to the inputfile forcing the profile to be fixed. But this does not yet propagate through into the final fit (ChangeAll)

As part of this I merged dfit, wfit and hfit. They are now a single function called ParamFit which can deal with fitting height width profile and dspacing of the peak.

These functions refineing the fit are now called from a loop the order of the components in which can easily be changed.

The symettry of the fourier series can also be set. This propagates through the code and into the output.

Sundry other changes have been made that are needed to work with the bulk of the changes.