Fix to rescaled colourmap
Plotting the fit residuals uses a modified colour map to keep the middle (0) a consistent colour. If the absolute minimum residual was greater than the biggest then the plot failed and crashed the code because the function produced an empty colour map.
This is now fixed.

The end of the function still contains the catch but commented out incase this is not fixed properly.