Remove **kwargs from execute
This forces options that are not used to issue and error.

The idea of this is that the propagation and tracking of the fits should be listed in the settings file and not passed to the function directly. Thus forcing every option that changes the fitting to be in the settings file for reproducibility.