Use multigeometry to make detector
Updated the functions after code from Wilson Crichton (ESRF). The code now makes a multidetector and then populates that. This also allows the class to have a 'detector' that is the correct item incase it is ever needed.

The code needs to be tidied up and further tested but this is an apparently working version.