Update DioptasFunctions to use common definitions
Reorganise, and rewrite DoiptasFunctions to use the common and mask functions. This also makes the masking and data functions clearer to understand -- there are many fewer methods to the class now.

Other files that are affected by the changes here are also updated. But this is not necessarily all the files that could be affected.