Update XY functions
Rewritten XYfunctions so that they use the same common methods as Dioptas and ESRFlvp classes.

Like the other classes, this has also cleaned up the masks and the plotting.

Plotting the data as a pseudo-image is still half developed and in the class. When time permits or it is necessary this will be finished and added to the plotting files.