Update to hdf5 keys
Make the information for processing hdf5 key lists more flexible and expandable. It is now dictionary based rather than varaible based. So everything has to be present and is harder confuse levels.

This is an inital attempt and should be easier to expand in future.

It is not as opaquely recursive as the previous method.