Example1 input tests
Edit the input files for example 1 to create test cases or examples for each of the processing features of the code.

The *input.py files do highlight different features of the code -- as listed in example function tests.rst file.

To do: same exersice for Example2.

The push also contains bug fixes for other parts of the code so that everthing runs.