Bug fix for initial series fitting
The initial series fit was being performed with a weighted fit, where the weights being used were the inverse of what they should have been.
How the fit deals with Nans was also changed.

A new function making sure that the errors are not stupidly large or small was also created. This is useful when the high --> 0 and the associated  error becomes --> <e-10.