Parallelisation fix
The ESRFlvp and MED detector classes could be parallelised for some (unknown to me) reason.

Removing the detector structure from the data_class (or a new data class is instantiated with no detector) allows the code to run in parallel. The data_class also has to be removed from the setttings class.

Both data and settings are now replicated using
calls to the classes of 'duplicate_without_dataclass'/'duplicate_without_detector'

New methods make a copy of the required data without copying the detector and calibrations. Neither of these are needed once the two theta and azimuth arrays are made.

This fix works for the ESRFlvp data class but not for the MED class.