reduce data size
Added a switch to reduce the size of the data sets. This is called by adding reduce_by = .. to the input file.
In testing it works for all the data types although as things stands fails because there are not enough points in the MgO regions to allow fitting. Changing n_min =21 to 7 does allow an outcome.