XYFunction fix.
XY functions cannot be assumed to be diffraction. If the conversion factor is False the conversion returns the same values it accepted.
This change has been made in the _AngleDispersie_common.conversion() function. Is this the best place for it? I am not sure. But for now it keeps the changes simple.