Metadata update
Update and restructuring of metadata functionality.
Now structured so that image metadata is added as a dictionary to the data_class on creation and then read via _metadata_common - get_metadata.
For the hdf5files the metadata is a dictionary of pointers that are needed to find the metadata in the hdf5 file.

Nothing is done with the metadata, it is just added to the data class. The next thing to do is to add it to the data structures and the write it to the json files.